+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-10T14:35:14.517Z",
+    "2026-02-11T00:35:30.193Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {